Changed the model model's default time and modified the test to have the same setup.
I've solved the problem of not being able to change the insert time. The solution is to change auto_now_add to timezone.now. The reason is because auto_now_add will ignore any input from the query. Due to the change in this, I've also committed a migration file. I've also changed the tests to have the same setup. Lastly, I removed the active flag from the query so that test_recovery_page_with_correct_email will have the two rows to proved that a new row was inserted.