Refactored all the forms into one.
Since all of the forms are of the same pattern, I've restructured all the forms to be under the same form class and then use this.props.mode to control the submission. This makes a cleaner code and makes it easier to change the look and feel of all forms at the same time.